requirements.txt
asn1crypto==0.24.0
Babel==2.5.3
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cryptography==2.1.4
Django==1.11.10
django-countries==5.1.1
django-phonenumber-field==2.0.0
djangorestframework==3.7.7
enum34==1.1.6
idna==2.6
ipaddress==1.0.19
phonenumberslite==8.9.0
Pillow==5.0.0
pycparser==2.18
PyJWT==1.6.0
pyOpenSSL==17.5.0
pytz==2018.3
requests==2.18.4
six==1.11.0
twilio==6.10.4
urllib3==1.22
